+++ packages/SequenceTranslator/src/apps/common/model/oligo-toolkit-package.ts
@@ -5,6 +5,8 @@ import * as DG from 'datagrok-api/dg';
 import {getMonomerLibHelper, IMonomerLibHelper} from '@datagrok-libraries/bio/src/monomer-works/monomer-utils';
 import {IMonomerLib, Monomer} from '@datagrok-libraries/bio/src/types';
 import {LoggerWrapper} from '@datagrok-libraries/bio/src/utils/logger';
+import {ISeqHelper} from '@datagrok-libraries/bio/src/utils/seq-helper';
+import {IHelmHelper} from '@datagrok-libraries/bio/src/helm/helm-helper';
 
 import {APP_NAME} from '../view/const';
 import {DEFAULT_LIB_FILENAME, FALLBACK_LIB_PATH} from './data-loader/const';
@@ -16,14 +18,18 @@ import {MonomerLibWrapper} from './monomer-lib/lib-wrapper';
 import {FormatConverter} from '../../translator/model/format-converter';
 import {FormatDetector} from './parsing-validation/format-detector';
 import {highlightInvalidSubsequence} from '../view/components/colored-input/input-painters';
-import {ISeqHelper} from '@datagrok-libraries/bio/src/utils/seq-helper';
 
 export class OligoToolkitPackage extends DG.Package implements ITranslationHelper {
-  private _seqHelper: ISeqHelper;
+
+  private _helmHelper: IHelmHelper;
+  public get helmHelper(): IHelmHelper {
+    if (!this._helmHelper)
+      throw new Error('Package SequenceTranslator .helmHelper is not initialized');
+    return this._helmHelper;
+  }
+
   public get seqHelper(): ISeqHelper {
-    if (!this._seqHelper)
-      throw new Error('Package SequenceTranslator .seqHelper is not initialized');
-    return this._seqHelper;
+    return this._helmHelper.seqHelper;
   }
 
   private _monomerLib?: IMonomerLib;
@@ -60,8 +66,8 @@ export class OligoToolkitPackage extends DG.Package implements ITranslationHelpe
     this._initPromise = initPromise;
   }
 
-  completeInit(seqHelper: ISeqHelper): void {
-    this._seqHelper = seqHelper;
+  completeInit(helmHelper: IHelmHelper): void {
+    this._helmHelper = helmHelper;
   }
 
   private initLibDataPromise?: Promise<void>;